+++ gfx/TSD_GFX.h
@@ -60,10 +60,11 @@ public:
 
   virtual void startWrite() = 0;
   virtual void endWrite() = 0;
-  virtual void writeAddrWindow(const int16_t x, const int16_t y, const int16_t w, const int16_t h) = 0;
+  virtual void writeAddrWindow(const int32_t x, const int32_t y, const int32_t w, const int32_t h) = 0;
   virtual void writeMDTBuffer(const uint8_t* buffer, const int32_t len) = 0;
   virtual void sendMDTColor1(const mdt_t c) = 0;
   virtual void sendMDTColor(const mdt_t c, const int32_t len);
-  virtual void drawClippedPixel(const int16_t x, const int16_t y, const rgb_t color);
-  virtual void drawClippedPixelRec(const int16_t x, const int16_t y, const int16_t w, const int16_t h, const rgb_t color);
+  virtual void drawClippedPixel(const int32_t x, const int32_t y, const rgb_t color);
+  virtual void drawClippedPixelRec(const int32_t x, const int32_t y, const int32_t w, const int32_t h, const rgb_t color);
+
 };
